devops_tools script updated

+++ vagrant/provision/devops_tools.sh
@@ -1,8 +1,28 @@
 #!/bin/bash
 
+## This is comment
+: '
+This script installs some DevOps tools
+Tested against: Ubuntu 18.04.5 LTS
+Latest update: 2020/12/26
+'
+
 echo "----------------------------------------------"
 echo "-----> Starting 'devops-tools' script <-------"
 echo "----------------------------------------------"
+
+## Help: https://stackoverflow.com/questions/1298066/check-if-an-apt-get-package-is-installed-and-then-install-it-if-its-not-on-linu
+
+## Define here versions and users
+
+DBEAVER_VERSION="7.3.1"
+DOCKER_COMPOSE_VERSION="1.27.4"
+GRADLE_VERSION="6.5.1"
+HELM_VERSION="helm-v3.4.2-linux-amd64.tar.gz"
+K9S_VERSION="v0.24.2"
+KUBECTL_VERSION=$(curl -s https://storage.googleapis.com/kubernetes-release/release/stable.txt)
+MINIKUBE_VERSION="v1.15.0"
+USER="vagrant"
 
 #####################
 ###### aws-cli ######
@@ -10,16 +30,24 @@ echo "----------------------------------------------"
 
 # Help: https://docs.aws.amazon.com/es_es/cli/latest/userguide/install-cliv2-linux.html
 
-echo "----> Installing aws-cli..."
+if [ ! -f /usr/local/bin/aws ]; then
 
-cd /tmp \
-&& curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip" \
-&& unzip awscliv2.zip \
-&& ./aws/install \
-&& rm -rf awscliv2.zip \
-&& echo "AWS-CLI Version --> $(aws --version)"
+  echo "----> Installing aws-cli..."
 
-echo "--> aws-cli successfully installed."
+  cd /tmp \
+  && curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip" \
+  && unzip awscliv2.zip \
+  && ./aws/install \
+  && rm -rf awscliv2.zip \
+  && echo "AWS-CLI Version --> $(aws --version)"
+
+  echo "--> aws-cli successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'aws' already installed!"
+
+fi
 
 ####################
 ###### az-cli ######
@@ -27,53 +55,83 @@ echo "--> aws-cli successfully installed."
 
 # Help: https://docs.microsoft.com/es-es/cli/azure/install-azure-cli?view=azure-cli-latest
 
-echo "----> Installing az-cli..."
+if ! dpkg -s azure-cli > /dev/null; then
 
-curl -sL https://aka.ms/InstallAzureCLIDeb | sudo bash
-echo "AZ-CLI Version --> $(az --version)"
+  echo "----> Installing az-cli..."
 
-echo "--> az-cli successfully installed."
+  curl -sL https://aka.ms/InstallAzureCLIDeb | sudo bash
+  echo "AZ-CLI Version --> $(az --version)"
+
+  echo "--> az-cli successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'azure-cli' already installed!"
+
+fi
 
 #####################
 ###### dbeaver ######
 #####################
 
-echo "----> Installing dbeaver..."
+if ! dpkg -s dbeaver-ce > /dev/null; then
 
-dbeaver_version="7.3.1"
+  echo "----> Installing dbeaver..."
 
-apt-get -y install default-jdk
-wget https://github.com/dbeaver/dbeaver/releases/download/${dbeaver_version}/dbeaver-ce_${dbeaver_version}_amd64.deb
-dpkg -i dbeaver-ce_${dbeaver_version}_amd64.deb
-rm -rf dbeaver-ce_${dbeaver_version}_amd64.deb
+  apt-get -y install default-jdk
+  wget https://github.com/dbeaver/dbeaver/releases/download/${DBEAVER_VERSION}/dbeaver-ce_${DBEAVER_VERSION}_amd64.deb
+  dpkg -i dbeaver-ce_${DBEAVER_VERSION}_amd64.deb
+  rm -rf dbeaver-ce_${DBEAVER_VERSION}_amd64.deb
 
-echo "--> dbeaver successfully installed."
+  echo "--> dbeaver successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'dbeaver-ce' already installed!"
+
+fi
 
 ####################
 ###### docker ######
 ####################
 
-echo "----> Installing docker."
+if ! dpkg -s docker-ce > /dev/null; then
 
-# Install docker
-apt-get update && apt-get install -y apt-transport-https ca-certificates gnupg-agent software-properties-common \
-&& curl -fsSL https://download.docker.com/linux/ubuntu/gpg | apt-key add - \
-&& add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" \
-&& apt-get update && apt-get install -y docker-ce docker-ce-cli containerd.io \
-&& sudo usermod -aG docker vagrant
+  echo "----> Installing docker."
 
-echo "--> docker successfully installed."
+  # Install docker
+  apt-get update && apt-get install -y apt-transport-https ca-certificates gnupg-agent software-properties-common \
+  && curl -fsSL https://download.docker.com/linux/ubuntu/gpg | apt-key add - \
+  && add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" \
+  && apt-get update && apt-get install -y docker-ce docker-ce-cli containerd.io \
+  && sudo usermod -aG docker ${USER}
+
+  echo "--> docker successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'docker-ce' already installed!"
+
+fi
 
 ############################
 ###### docker-compose ######
 ############################
 
-echo "----> Installing docker-compose"
+if [ ! -f /usr/local/bin/docker-compose ]; then
 
-curl -L "https://github.com/docker/compose/releases/download/1.25.4/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose \
-&& chmod +x /usr/local/bin/docker-compose
+  echo "----> Installing docker-compose"
 
-echo "--> docker-compose successfully installed."
+  curl -L "https://github.com/docker/compose/releases/download/${DOCKER_COMPOSE_VERSION}/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose \
+  && chmod +x /usr/local/bin/docker-compose
+
+  echo "--> docker-compose successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'docker-compose' already installed!"
+
+fi
 
 ##################
 ##### gradle #####
@@ -81,20 +139,26 @@ echo "--> docker-compose successfully installed."
 
 # Download page: https://gradle.org/install/
 
-gradle_version="6.5.1"
+if [ ! -d /opt/gradle ]; then
 
-echo "----> Installing gradle"
+  echo "----> Installing gradle"
 
-mkdir /opt/gradle
-cd /opt/gradle
-wget https://services.gradle.org/distributions/gradle-${gradle_version}-bin.zip
-unzip -d /opt/gradle gradle-${gradle_version}-bin.zip
-rm -rf gradle-${gradle_version}-bin.zip
-# Permanently add gradle to path
-runuser -l vagrant -c "echo "PATH="$PATH:/opt/gradle/gradle-${gradle_version}/bin"" >> /home/vagrant/.profile"
-gradle -v
+  mkdir /opt/gradle
+  cd /opt/gradle
+  wget https://services.gradle.org/distributions/gradle-${GRADLE_VERSION}-bin.zip
+  unzip -d /opt/gradle gradle-${GRADLE_VERSION}-bin.zip
+  rm -rf gradle-${GRADLE_VERSION}-bin.zip
+  # Permanently add gradle to path
+  runuser -l vagrant -c "echo "PATH="$PATH:/opt/gradle/gradle-${GRADLE_VERSION}/bin"" >> /home/vagrant/.profile"
+  gradle -v
 
-echo "--> gradle successfully installed."
+  echo "--> gradle successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'gradle' already installed!"
+
+fi
 
 ####################
 ####### helm #######
@@ -102,18 +166,24 @@ echo "--> gradle successfully installed."
 
 # Download page: https://github.com/helm/helm/releases
 
-helm_version="helm-v3.4.2-linux-amd64.tar.gz"
+if [ ! -f /usr/local/bin/helm ]; then
 
-echo "----> Installing helm."
+  echo "----> Installing helm."
 
-cd /tmp
-wget https://get.helm.sh/$helm_version
-tar -zxvf $helm_version
-cp linux-amd64/helm /usr/local/bin/helm
-rm -rf $helm_version linux-amd64/
-echo "Helm Version --> $(helm version)"
+  cd /tmp
+  wget https://get.helm.sh/$HELM_VERSION
+  tar -zxvf $HELM_VERSION
+  cp linux-amd64/helm /usr/local/bin/helm
+  rm -rf $HELM_VERSION linux-amd64/
+  echo "Helm Version --> $(helm version)"
 
-echo "--> helm successfully installed."
+  echo "--> helm successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'helm' already installed!"
+
+fi
 
 ####################
 ##### intellij #####
@@ -133,17 +203,23 @@ echo "--> intellij successfully installed."
 ###### k9s ######
 #################
 
-K9S_VERSION="v0.24.2"
+if [ ! -f /usr/local/bin/k9s ]; then
 
-cd /tmp
-wget https://github.com/derailed/k9s/releases/download/$K9S_VERSION/k9s_Linux_x86_64.tar.gz
-tar -zxvf k9s_Linux_x86_64.tar.gz
-cp k9s /usr/local/bin/k9s
+  cd /tmp
+  wget https://github.com/derailed/k9s/releases/download/$K9S_VERSION/k9s_Linux_x86_64.tar.gz
+  tar -zxvf k9s_Linux_x86_64.tar.gz
+  cp k9s /usr/local/bin/k9s
 
-rm -rf k9s_Linux_x86_64.tar.gz k9s LICENCE README.md
-echo "k9s Version --> $(k9s version)"
+  rm -rf k9s_Linux_x86_64.tar.gz k9s LICENCE README.md
+  echo "k9s Version --> $(k9s version)"
 
-echo "--> k9s successfully installed."
+  echo "--> k9s successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'k9s' already installed!"
+
+fi
 
 #####################
 ###### kubectl ######
@@ -151,30 +227,44 @@ echo "--> k9s successfully installed."
 
 # Download page: https://github.com/kubernetes/kubectl/releases
 
-kubectl_version="v1.17.4"
+if [ ! -f /usr/local/bin/kubectl ]; then
 
-echo "----> Installing kubectl."
+  echo "----> Installing kubectl."
 
-cd /tmp
-curl -Lo kubectl https://storage.googleapis.com/kubernetes-release/release/$kubectl_version/bin/linux/amd64/kubectl
-chmod +x kubectl
-sudo mv ./kubectl /usr/local/bin/kubectl
-kubectl version
-echo "kubectl Version --> $(kubectl version)"
+  cd /tmp
+  curl -Lo kubectl https://storage.googleapis.com/kubernetes-release/release/$KUBECTL_VERSION/bin/linux/amd64/kubectl
+  chmod +x kubectl
+  sudo mv ./kubectl /usr/local/bin/kubectl
+  kubectl version
+  echo "kubectl Version --> $(kubectl version)"
 
-echo "--> kubectl successfully installed."
+  echo "--> kubectl successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'kubectl' already installed!"
+
+fi
 
 ###################
 ###### maven ######
 ###################
 
-echo "----> Installing mvn."
+if ! dpkg -s maven > /dev/null; then
 
-apt-get update
-apt-get install -y maven
-echo "maven Version --> $(mvn -version)"
+  echo "----> Installing mvn."
 
-echo "--> mvn successfully installed."
+  apt-get update
+  apt-get install -y maven
+  echo "maven Version --> $(mvn -version)"
+
+  echo "--> mvn successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'maven' already installed!"
+
+fi
 
 ######################
 ###### minikube ######
@@ -182,17 +272,23 @@ echo "--> mvn successfully installed."
 
 # Download page: https://github.com/kubernetes/minikube/releases/
 
-minikube_version="v1.16.0"
+if [ ! -f /usr/local/bin/minikube ]; then
 
-echo "----> Installing minikube."
+  echo "----> Installing minikube."
 
-cd /tmp
-curl -Lo minikube https://github.com/kubernetes/minikube/releases/download/$minikube_version/minikube-linux-amd64
-chmod +x minikube
-cp minikube /usr/local/bin && rm minikube
-echo "Minikube Version --> $(minikube version)"
+  cd /tmp
+  curl -Lo minikube https://github.com/kubernetes/minikube/releases/download/$MINIKUBE_VERSION/minikube-linux-amd64
+  chmod +x minikube
+  cp minikube /usr/local/bin && rm minikube
+  echo "Minikube Version --> $(minikube version)"
 
-echo "--> Minikube successfully installed."
+  echo "--> Minikube successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'minikube' already installed!"
+
+fi
 
 ######################
 ###### skaffold ######
@@ -200,43 +296,59 @@ echo "--> Minikube successfully installed."
 
 # Download page: https://skaffold.dev/docs/install/
 
-echo "--> Installing skaffold."
+if [ ! -f /usr/local/bin/skaffold ]; then
 
-cd /tmp
-curl -Lo skaffold https://storage.googleapis.com/skaffold/releases/latest/skaffold-linux-amd64
-install skaffold /usr/local/bin/
-rm skaffold
-echo "skaffold Version --> $(skaffold version)"
+  echo "--> Installing skaffold."
 
-echo "--> skaffold successfully installed."
+  cd /tmp
+  curl -Lo skaffold https://storage.googleapis.com/skaffold/releases/latest/skaffold-linux-amd64
+  install skaffold /usr/local/bin/
+  rm skaffold
+  echo "skaffold Version --> $(skaffold version)"
+
+  echo "--> skaffold successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'skaffold' already installed!"
+
+fi
 
 #####################
 ###### vs-code ######
 #####################
 
-echo "----> Installing vs-code..."
+if ! dpkg -s code > /dev/null; then
 
-# Download and install vs-code
-apt-get update
-apt-get install -y software-properties-common apt-transport-https wget
-wget -q https://packages.microsoft.com/keys/microsoft.asc -O- | apt-key add -
-add-apt-repository "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main"
-apt-get update
-apt-get install -y code
+  echo "----> Installing vs-code..."
 
-# Fix: https://askubuntu.com/questions/760896/how-can-i-fix-apt-error-w-target-packages-is-configured-multiple-times
-rm -rf /etc/apt/sources.list.d/vscode.list
+  # Download and install vs-code
+  apt-get update
+  apt-get install -y software-properties-common apt-transport-https wget
+  wget -q https://packages.microsoft.com/keys/microsoft.asc -O- | apt-key add -
+  add-apt-repository "deb [arch=amd64] https://packages.microsoft.com/repos/vscode stable main"
+  apt-get update
+  apt-get install -y code
 
-# Install plugins from cli - https://code.visualstudio.com/docs/editor/extension-gallery
-plugins=( "donjayamanne.githistory" "eamodio.gitlens" "hashicorp.terraform" "ivory-lab.jenkinsfile-support" "korekontrol.saltstack" "marcostazi.VS-code-vagrantfile" "mhutchie.git-graph" "mhutchie.git-graph" "ms-kubernetes-tools.vscode-kubernetes-tools" "ms-python.python" "redhat.vscode-yaml" "vscode-icons-team.vscode-icons" "vscoss.vscode-ansible" "wholroyd.jinja" "yzhang.markdown-all-in-one" "zhangciwu.swig-tpl")
+  # Fix: https://askubuntu.com/questions/760896/how-can-i-fix-apt-error-w-target-packages-is-configured-multiple-times
+  rm -rf /etc/apt/sources.list.d/vscode.list
 
-for i in "${plugins[@]}"
-do
+  # Install plugins from cli - https://code.visualstudio.com/docs/editor/extension-gallery
+  plugins=( "donjayamanne.githistory" "eamodio.gitlens" "hashicorp.terraform" "ivory-lab.jenkinsfile-support" "korekontrol.saltstack" "marcostazi.VS-code-vagrantfile" "mhutchie.git-graph" "mhutchie.git-graph" "ms-azuretools.vscode-docker" "ms-kubernetes-tools.vscode-kubernetes-tools" "ms-python.python" "redhat.vscode-yaml" "vscode-icons-team.vscode-icons" "vscoss.vscode-ansible" "wholroyd.jinja" "yzhang.markdown-all-in-one" "zhangciwu.swig-tpl")
+
+  for i in "${plugins[@]}"
+  do
   echo "--> Installing plugin $i"
-  runuser -l vagrant -c "code --install-extension $i"
-done
+  runuser -l ${USER} -c "code --install-extension $i"
+  done
 
-echo "--> vs-code successfully installed."
+  echo "--> vs-code successfully installed."
+
+else
+
+  echo "--> PACKAGE: 'vs-code' already installed!"
+
+fi
 
 ########################
 ###### favourites ######
